Components seperate in to sub components to easier handlig and debuggin
